pcap_breakloop(3PCAP): describe the shiny new world. [skip ci]
Note that, on Linux and Windows, pcap_breakloop() attempts to force a
wakeup on the thread reading the packets.

Note all the hilarity that ensues from that; this is one of the areas
where capture-mechanism differences are not quite so easily papered
over.

(These capture mechanisms were all pretty much designed for a program
that reads captures until it's told to stop, and then may or may not
process packets that arrived after that point.  Making it available
through libpcap has opened up packet capturing to a zillion developers,
many of whom are doing different types of capture programs with
different requirements, and stumbling over all these problems.  Perhaps
it's time for a rethink.

We won't even mention the use of the capture mechanisms for purposes
*other* than packet capture, such as implementing various protocols atop
the link-layer; libpcap's API is not the best fit for that - a
corss-platform "link-layer sockets" library, possibly using a
*different* OS mechanism than libpcap on some platforms, might be handy
there.)

(cherry picked from commit 6795eae85bd54980c78c085591bafe7436bae174)