autoconf: Add autogen.sh, remove configure and config.h.in
Put autoconf-generated files in the release tarball.

Remove the corresponding task from the TODO file.

The minimum required version of autoconf is currently 2.69.

If version 2.69 or later is already installed and there is no
autoconf default, it may be necessary to set the AUTORECONF
environment variable to enable the one to use, like:
AUTORECONF=autoreconf-2.69 ./autogen.sh
or
AUTORECONF=autoreconf-2.71 ./autogen.sh