configure: properly report errors from --cflags/--libs pkg-config calls.
That will do a better job of reporting errors such as the ones that
showed up with pre-0.17 versions of pkg-config, so the problem is more
obvious.

While we're at it, rename PKG_CHECK_MODULES to PKG_CHECK_MODULE, and
rename some argument names, to reflect that PKG_CHECK_MODULE is not set
up to check more than one module, now that it sets variables for the
module.