configure: fix handling of Haiku -D_BSD_SOURCE and -lbsd flags.
Add -D_BSD_SOURCE to V_DEFS, as we leave CFLAGS alone and add additional
-D flags to V_DEFS.

Don't save the check state before checking for special Haiku flags, as
we'll be adding libraries to LIBS; instead, save it after we do the
check.

This ensures that -D_BSD_SOURCE is passed to the compiler and -lbsd is
passed to the linker in the resulting Makefile.