Simplify conditional branching in diag-control.h.
Since diagnostic macros do not depend on each other, use the same
approach as in tcpslice: make one or more attempts to define a
diagnostic macro depending on conditions of varying complexity; if after
all the busy branching the macro is not yet defined, define it to an
empty string once.  This way the branches are simpler to rearrange when
required, and the file has a manifest of the macro names it defines.