Document pcap-config better and test it. [skip ci]
Reformat "pcap-config --help" message for readability and include
--cflags.  Rewrite the man page to use the same format in the synopsis,
make the options description a standalone section, document the recently
implemented --help, --version and --static-pcap-only options, add a
section to document the exit status and add a reference to
pkg-config(1).  In build.sh try various invocations of pcap-config to
make sure it still recognizes the documented options and to aid in
troubleshooting of failed builds.