autoconf: Add the option to print functions and files names
This should help some debugging or learning processes in combination
with tcpdump (or tcpslice).

Usage:
$ ./configure --enable-instrument-functions

This is a preparatory step that also requires to use
$ ./configure --enable-instrument-functions
in tcpdump (or tcpslice).

Generate instrumentation calls for entry and exit to functions.
Just after function entry and just before function exit, the
profiling functions are called and print the function names with
indentation and call level.
If entering in a function, print also the calling function name with
file name and line number. There may be a small shift in the line number.

For the record, in tcpdump (or tcpslice), the following options are
available:

To configure the printing of only the global functions names:
$ make instrument_global

To go back to print all the functions names:
$ make instrument_all

To print nothing, like with no instrumentation:
$ make instrument_off

[skip ci]