npf: clean up handling of errors from setting promiscuous mode.
Handle both ERROR_NOT_SUPPORTED and "NDIS_STATUS_NOT_SUPPORTED with the
Customer bit set" as indications that an attempt to set promiscuous mode
was rejected because the adapter doesn't support it, in case the setting
of the Customer bit is dropped from the Npcap NPF driver in the future.

While we're at it, shuffle some #defines for status codes to put them
together and update comments.