Make i, n, index, and max_stmts unsigned integers
This fixes compiler warnings regarding signedness
and avoids casting.

maxval should be an unsigned integer because
slength, the result of which is added to this number,
returns an unsigned value.

In addition, opt_state->maxval being an unsigned int
that is assigned via multiplying max_stmts is also a
good case for max_stmts being unsigned.

Finally, count_blocks should be unsigned since the
minimum number is 0, and it assigns to n, which "i"
is compared to, and is used for pointer arithmetic.