Remove redundant assigning of p1 to prev
The first assignment of p1 to prev is redundant, as the assignment to prev from p1 happens after the if statement anyway.

This was found by clang-tidy emitting the following warning: "Value stored to 'prev' is never read", as prev is reassigned to the same expression once outside the if-statement.