Rename doc/README.Win32.md to doc/README.windows.md.
This gives it a name similar to the other README.{operatingsystem}
files.

It also reflects that 1) 16-bit Windows is now so old and tired that
there's no need to use a name that specifically mentions a non-16-bit
version of Windows to indicates that we don't support it and 2) there's
now 64-bit Windows, so "Win32" is an out-of-date name.