Fail if packet capture support cannot be found.
On Windows, tell the user that they need to install Npcap or WinPcap,
and the respective SDK, and tell CMake where the SDK is located.  On
UN*X, tell them to check INSTALL.md.

In addition, tell the user to set the capture type to "null" when
configuring if they just want a version of libpcap that can read capture
files but can't do live captures.

That way, the user won't be surprised at run time by a libpcap that
fails with "live packet capture not supported on this system" if they
try to capture traffic.

Fixes #1016 (in the sense that configuration stops if a capture
mechanism can't be found, and tells the user what they need to do).