appveyor: we're installing the 1.13 Npcap SDK, use it, not the 1.12 SDK.
One advantage of d7f44c5f8328a3979286ed2715f900a3a556001f is that it
revealed that we were installing the 1.13 Npcap SDK and trying to
configure with the 1.12 SDK; now that doing so fails, the build fails.

Fix that, so we use the same SDK as we install.