Fix universal builds.
Move the stuff to figure out for what architectures to build before we
do anything with rpcapd.

Have that stuff decide both what architectures to build libraries for
and what architectures to build executables for.  Call the executables
list OSX_EXECUTABLE_ARCHITECTURES rather than OSX_PROGRAM_ARCHITECTURES,
to match the terminology used in the names of the "add a target"
commands.

Remove the not-updated-for-ARM stuff in rpcapd's CMakeLists.txt, and
just use the architecture list we picked in the main CMakeLists.txt.

Remove some debugging messages.