bpf: report {non-existent zone}/{interface} errors appropriately.
"Invalid argument" doesn't indicate that the problem is that the zone in
question doesn't exist and "File name too long" doesn't indicate that
the problem is that the zone in question doesn't exist because it can't
possibly exist because zone names that long are not allowed.

Report both as "no such device on which to capture" rather than just as
a generic error.