bpf: move handling of zones into bpf_bind().
That's where the binding is done, so that's where we need to get a zone
ID and interface name within that zone from a device name.

This is based on pull request #1202, but reports errors correctly, and
works even when built on older versions of Solaris 11 without zone
support in BIOCSETLIF.

Fix/update some comments while we're at it.