More temporary debugging stuff. [skip ci]
Remove all the pre-VS 2019 AppVeyor builds, as they don't have the
problematic OpenSSL.

In the build script, look for any opensslv.h* file under the top-level
include directory of the problematic SSL; that's the file that sets the
OPENSSL_VERSION variabls.

Add [skip ci] in the hopes of provoking only AppVeyor (and OpenCSW?) to
build this commit.