AppVeyor: disable remote capture with VS 2019.
The OpenSSL library on the current AppVeyor Visual Studio 2019 images
has a weird opensslv.h that claims its 1.0.2, even though it's 3.0.
This causes...  problems.

For now, we disable the remote capture build there.

Back out the debugging stuff and the attempt to fix it in sslutils.c -
the weird opensslv.h causes that not to work.