Cirrus CI: use curl -o and -O, quote AirPcap SDK URL.
Use "curl -O" to put the download in a file in the current directory
with the same name as the last component of the URL.

Use "curl -o" to put the download in a file in the current directory
with a specified name.

Quote the URL for the AirPcap SDK, just in case it's being treated
weirdly a Windows argument-list parser.