Cirrus CI: some Windows tweaks.
Move env: before install_script:, as it should affect both.

Add MSYSTEM: MINGW64, so that we do a MinGW64 build rather than an MSYS
build, as we're building native Windows software.

Do some directory searches to see what binaries come with Git and where
uname came from (not from MSYS2, as we weren't installing it when we
first tested it).

Try using a Windows path to build_matrix.sh.