Cirrus CI: dump out the /etc/profile and .profile files.
Cygwin's shell (which is what's used in MSYS2 and git-bash, the latter
of which appears to be installed by Cirrus CI) screws up massively when
constructing the UN*X path from the Windows PATH if any elements in the
Windows PATH have blanks - such as anything under C:\Program Files.

This causes the UN*X path to contain mangled versions of the git-bash
executable directories, so that, for example,

     C:\Program Files\Git\usr\bin

is turned into

	/usr/bin

rather than into

	/c/Windows/Program Files/Git/usr/bin

(which needs the blanks to be escaped *if you type it on the command
line*, but is *perfectly OK with the spaces in it* as part of an
environment variable!).

As a result, none of the git-bash commands are found.

This... is not good.

See if the Windows path is imported in the global or personal shell
profiles.