Cirrus CI: dump directories under C:\Program Files\Git selectively.
There's too much stuff under there; look at the directories that appear
in the Windows PATH.  Look at the new stuff in the UN*X path as well; it
might be that C:\Program Files\Git is the MinGW/MSYS2 root, so the UN*X
PATH isn't getting mangled.  (And it may be that the failure to find
cmake may be that cmake isn't installed by default, so we'd have to
choco it or something such as that.  It'd be nice if it were installed
as part of Visual Studio in the containers.)