Redo the availability macros.
Apple appears to have tweaked their pcap/funcattrs.h to do the right
thing at some point, so 1) there's no need to urge them to do so any
more and 2) they only seem to export libpcap APIs in macOS, so redo the
PCAP_AVAILABLE() stuff to be more like what they've done.

We remove the include of <Availability.h>, as we don't need it ourselves
(we don't do the availability stuff in our releases, for reasons given
in the comment before the availability stuff), and Apple uses
<os/availability.h> - we leave it up to them to modify the header, just
as they do with, for example, ZLib's zlib.h.

(Many thanks to the Apple people who picked this up!)