instrument functions: Use two loop indexes with reduced scopes
The first is a long, comparison with the long 'symcount'.
The second is a int, comparison with the int 'profile_func_level'.

Moreover:
Add a cast to silence the following warning:
./instrument-functions.c:143:31: warning: implicit conversion changes
signedness: 'long' to 'unsigned long' [-Wsign-conversion]
        symtab = (asymbol **)malloc(symsize);
                             ~~~~~~ ^~~~~~~
(symsize is a long as return type of bfd_get_symtab_upper_bound().)

(Same as in tcpdump.)

[skip ci]