Fix availability macros to handle macOS Sonoma.
The remote capture APIs are, in 1.9 and 1.10, usually only available in
the library if the library was built with remote capture enabled.

macOS Sonoma provides stub versions of those routine, which return an
error.  This means that we need a separate availability indicator macro
for those routines, for reasons explained in a comment in
pcap/funcattrs.h.

(See https://gitlab.com/wireshark/wireshark/-/issues/19436 for an
instance of the problem that this causes.)

(cherry picked from commit 9de890f43c722af79848d532c1a38b035f578e2d)