pcap-linux: fix a file descriptor link.
If we're not going to be in non-blocking mode, open the event FD in the
activate rourinte; we don't need it until we're going to activate it for
capturing.

This means we don't need to close it if a non-activated handle is being
closed.

Clean up some comments, etc. while we're at it.

Fixes issue #1233.