pcap-linux: squelch a warning.
Unfortunaely, pcap_breakloop() doesn't return a status, so we can't
report an error.  Thus, we ignore the return value of write() on the
event FD, and GCC doesn't accept casting that return value to void as a
way of saying "I'm deliberately ignoring this return value, don't
complain about it".

Suppress that warning with a pragma.

(cherry picked from commit 079653047fa480c0337d8bb2094769067a525008)