optimizer: move more "optimizer loop detection" code.
Shuffle most of it to the end of the block, to keep all the other stuff
together and separate the "optimizer loop detection" stuff.

Move one to the *beginning* of the block, as it's clearing the
non_branch_movement_performed flag before any work is done, and
eliminate one, as it's unnecessary given that it clears that flag and
given that the structure it's in was memset() to 0 just before that.