socketcan: fix up CAN/CAN FD support and add CAN XL support.
Add code to format the CAN XL header in a byte-order-independent
fashion; use little-endian, to match, as best as possible, what existing
captures (most of which were probably done on little-endian machines)
have been done.

Always look at the protocol type and forcibly set or clear the appropriate
flags in the header, to increase the chances that they're set correctly.