Hurd: Count and timestamp packets better.
The expected behaviour of "tcpdump -c 10 tcp" is to exit after capturing
10 TCP packets no matter how many other packets were on the wire, which
works on other OSes.  On GNU/Hurd the actual behaviour is exiting after
seeing any 10 packets on the wire no matter how many were TCP, if any.
This is because pcap_read_hurd() always returns 1 for a packet, and it
receives all packets because the filtering occurs in userland.  To fix
this, return 1 iff the packet matched the filter.

While at it, obtain the packet timestamp ASAP to make it a bit more
accurate.