socketcan: put the first field of a CAN XL header in big-endian order.
Old CAN XL-unaware versions of libpcap did so, as they thought it was
the CAN ID field.  So that readers of LINKTYPE_CAN_SOCKETCAN files can
handle those captures and fixed captures, we make that field big-endian.
We continue to make the other two multi-byte fields of the CAN XL header
little-endian, as those old versions of libpcap didn't change tham, and
most if not all CAN XL captures were done on little-endian platforms.