Haiku: Report non-existent devices correctly. [skip ci]
The only means to return an error from pcapint_create_interface() is to
return NULL, so it is not the right place for a code that needs to
report specific errors by specific error codes.  Move most of the device
initialization code into pcap_activate_haiku() and when the device does
not exist, return PCAP_ERROR_NO_SUCH_DEVICE as it ought to be done.

tcpdump -D
1./dev/net/ipro1000/0 [Up, Connection status unknown]
2.loop [Up, Loopback]

Before:

> ./tcpdump -i 1 -J
tcpdump: Interface "1" does not exist.

After:

> ./tcpdump -i 1 -J
Time stamp types for 1 (use option -j to set):
  host (Host)