Haiku: Address assorted remaining issues. [skip ci]
In pcap_read_haiku() call recvfrom() without the sockaddr outputs,
handle the wire and the captured lengths correctly regardless if the
packet was truncated or not and in case the wire length does not fit
into 32 bits just drop the packet early instead of returning an error.
In pcap_cleanup_haiku() use pcapint_cleanup_live_common() instead of
trying to duplicate it.  Add a few comments to document difficulty and
usefulness of various potential improvements.  Do not use pcap_t.offset.