Fix Autoconf builds with illumos make. [skip appveyor]
My recent commit 087bc2e had broken building of the test programs using
Autoconf and illumos make, but not the illumos CI build, which has been
using GNU Make, supposedly because illumos make did not support parallel
builds when the CI build was set up.

As a matter of fact, it supports parallel builds now so long as there is
a space between the "-j" flag and its argument, so let's fix the
activatetest target before switching the CI job to illumos make.  While
at it, use "$@" in all other targets to reduce unnecessary duplication.