Fix propagation of cc_werr_cflags() output.
For testprogs/filtertest.c Sun C compiler versions 5.14 and 5.15 have
been generating three warnings about unreachable code, but that on its
own causes neither Autoconf nor CMake legs to fail when LIBPCAP_TAINTED
is set to "no".  The same applies to testprogs/fuzz/fuzz_filter.c, CMake
and GCC/Clang on FreeBSD, macOS and Linux.  The problem is, the result
of cc_werr_cflags() has no effect on testprogs/, which is not obvious so
long as the compiler does not generate warnings there.

To debug this, in cc_werr_cflags() return two options for Sun C and see
it breaking the detection of warnings entirely in the CMake leg:

-- Checking C compiler flag -errwarn=%all -errtags=yes
-- Performing Test errwarn-%all -errtags-yes
-- Performing Test errwarn-%all -errtags-yes - Failed
-- Added extra compile options (-errwarn=%all -errtags=yes)
[...]
".../libpcap/testprogs/filtertest.c", line 258: warning: statement not
  reached
[ 96%] Linking C executable ../run/filtertest
[ 96%] Built target filtertest

The original intent in commit f9403bf was to allow more than one C
compiler option in EXTRA_CFLAGS.  Then commits 29d7856 and 815b367
restored compatibility with CMake 2.8.6 -- the minimum supported version
at the time -- but introduced another bug, which was not obvious so long
as EXTRA_CFLAGS defined at most one option.  Since the minimum is at
2.8.12 now and add_compile_options() is available, use it as originally
intended:

-- Added extra compile options (-errwarn=%all -errtags=yes)
[...]
".../libpcap/testprogs/filtertest.c", line 258: statement not reached
  (E_STATEMENT_NOT_REACHED)
cc: acomp failed for .../libpcap/testprogs/filtertest.c
*** Error code 2

A side effect of this is that the CMake leg now applies EXTRA_CFLAGS to
all directories.  In the Autoconf leg pass CFLAGS on the command line.
Now all remaining warnings fail the build as expected.