Print MAC addresses in findalldevstest. [skip appveyor]
If AF_PACKET or AF_LINK is defined, include the headers and enable the
code required to print at least the basic properties of the address
family.  If the type and the length of the address match Ethernet, print
the [MAC] address.  Mask the last three octets of MAC addresses unless
the UNMASK_MAC_ADDRESSES environment variable is set to "yes".