Improve the description of caplen and len.
Don't say "off wire"; once upon a time, the data provided when capturing
corresponded to the octets "on the wire" or "on the air" at the data
link layer, but, now, the data provided when capturing might:

* include a metadata header that doesn't show up on the wire/air;
* be the result of ATM AAL5 reassembly rather than just be raw cells;
* represent a USB transaction as seen by the host rather than raw USB
  traffic;
* include padding between chunks of data in an isochronous USB transfer;
* etc.

And even in the old days, the data might, or might not, include the CRC
at the end of a packet.

Don't encourage people to think that the len field corresponds to the
length on the wire, as it might not.