Make pcap_strerror() more thread-safe.
Drop the checks for strerror() and the code block in pcap_strerror()
that resorts to sys_nerr and sys_errlist[] because HAVE_STRERROR is
always defined because every supported OS implements strerror(), which
is in POSIX.1-2001.

Move the GNU- and POSIX-specific strerror_r() code blocks from
pcapint_vfmt_errmsg_for_errno() to pcap_strerror() to make the latter
thread-safe when strerror_r() is available.  If POSIX strerror_r()
fails, try to fill the buffer with a useful string even if the failure
cannot be interpreted.

Update comments to explain the flow of conditionals with regard to
various OSes.