socketcan: make CAN CC/FD/XL frame handling less invasive
The different CAN frame types are defined by Linux SLL_P types in the
sll_protocol field and the length of the frame.

LINUX_SLL_P_CANXL:
The frame length for CAN XL can be 12 + 1 to 12 + 2048 (13 .. 2060) byte.
Additionally the CANXL_XLF flag has to be set.

LINUX_SLL_P_CAN and LINUX_SLL_P_CANFD:
The frame length for CAN CC is 16 byte and for CAN FD it is 72 byte.
Additionally the CANXL_XLF flag is cleared in CAN CC/FD frames.

Reduce the modification and 'sanitizing' efforts only to the above
requirements to be robust against incompliant content which might be
provided via PACKET sockets but do not manipulate further content.

This helps to identify uninitialized data and to support new features
in formerly reserved structure elements without changing the libpcap.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>