usb-linux: do a sanity check on the number of descriptors.
The Linux binary USB monitor code limits the number of isochronous
descriptors to 128; if the number in the file is larger than that,
either 1) the file's been damaged or 2) the file was produced after the
number was raised in the kernel.

In case 1), the number can't be trusted, so don't rely on it to attempt
to fix the original length field in the pcap or pcapng header.

In case 2), the system was probably running a version of libpcap that
didn't miscalculate the original length, so it probably doesn't need to
be fixed.

This avoids the possibility of the product of the number of descriptors
and the size of descriptors won't overflow an unsigned 32-bit integer.