usb-linux: reorganize the isochronous length handling code.
Put useful inline functions, used both by the capture code and the
savefile length fixup code, into pcap-usb-linux-common.h; some of those
came from, or are based on routines in, pcap-usb-linux-common.c, which
we remove.

When capturing, all we need to do is check whether a message with data
is a "this is complete" incoming isochronous transfer event and, if so,
use the routine to calculate its original length.

Pull the routine to fix up the original length into
pcapint_post_process().

When reading savefiles, we need to do additional checks to guess whether
the length needs fixing and to determine whether the number of
descriptors is sane and whether we have all the descriptors; do that in
pcapint_post_process().

Add some additional checks.

Call the "len" field of pcap's metadata header the "original length"
rather than the "on-the-wire" length; that matches the terminology from
the current versions of the pcap and pcapng specs, and doesn't imply
that the length is the length of the packet on the transmission medium,
as that's not always the case - for USB, it's computed based on
information from the endpoint on which we're capturing, and may also
include padding between isochronous data chunks.

(cherry picked from commit 2d12e865c43dddf064b50b1eba276e49d750fe29)