The time stamp in pcap packet headers has two unsigned 32-bit values.
When writing a pcap packet record, cast the seconds and microseconds
values to bpf_u_int32, as that's the type of the values in the record
header.  Update the sell-by date of pcap files; making them unsigned
means pcap files don't have a Y2.038K problem, but it does have a
Y2.106K problem.