CMake: Remove unnecessary header checks.
The checks for <inttypes.h> and <stdint.h> always succeed in a C99
environment, thus HAVE_INTTYPES_H and HAVE_STDINT_H are always defined,
but that is irrelevant because the code includes the headers regardless
of the macros.  Likewise, the check for <bitypes.h> always fails because
the header is supposed to be <sys/bitypes.h>, but that is irrelevant
because CMake does not use the test result and the code does not use the
header since commit a8f73cd.

While at it, lose the no-op HAVE_STDLIB_H and HAVE_STRING_H too.