CMake: remove some unchecked-for HAVE_ values.
Those were copied and pasted from config.h.in, which defined them
because autoconf inserts tests for them in all generated scripts, even
if nobody needs the test.

We don't test for them, so we might as well just leave them out of
cmakeconfig.h.in.

While we're at it, change a comment to match config.h.in, and fix the
capitalization of Npcap in a comment.