Rename {LINKTYPE,DLT}_IPMB_LINUX to {LINKTYPE,DLT}_I2C_LINUX.
The link-layer header just carries I2C information, and the initially
proposed names had I2C (first LINKTYPE_I2C, and then LINKTYPE_I2C_LINUX
as there was no guarantee that it would be the only I2C format used).

However, as using it as a general I2C header would require that packet
analyers have a knob to select what protocol the payload is using, and
it was requested for IPMB-over-I2C, so we decided to call it
LINKTYPE_IPMB_LINUX.

But it's now also used for HDMI Display Data Channel traffic, which also
runs over I2C, so it's no longer guaranteed to be IPMB traffic - and,
sure enough, Wireshark now has a "Decode As..." option to indicate what
protocol is being used.

So let's just go back to calling it by a name that reflects how it's
being used in practice, and leave DLT_IPMB_LINUX around as an alternate
name, for backwards source compatibility.